requirements.txt
anthropic
asyncpg
beautifulsoup4
langchain
langchain-community
langchain-core
langchain-text-splitters
numpy
pandas
pgvector
psycopg2-binary
pydantic
pydantic-ai
pydantic-settings
python-dotenv
requests
sqlalchemy
tiktoken
openai
fastapi
uvicorn[standard]
python-multipart
rich
groq
pytest
pytest-asyncio

# Enhanced RAG dependencies
nltk>=3.8
cohere>=4.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
pyyaml>=6.0